10-robust-testing(pytest)/pyproject.toml
[tool.poetry]
name = "10-robust-testing(pytest)"
version = "0.1.0"
description = ""
authors = ["ahnsv <ahnsv@bc.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.105.0"}
sqlmodel = "^0.0.14"
pydantic-settings = "^2.1.0"
loguru = "^0.7.2"
transformers = "4.35.2"
pytorch-lightning = "2.1.2"
scikit-learn = "^1.3.2"
nltk = "^3.8.1"
pillow = "^10.2.0"
torchvision = "^0.16.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
